[#39] Convert mock to actual component TypeCascade
